Added code for new LCD panels: ERC12864-14 and ERC24064_1 along with changes in the makefile to choose the ERC lcd type at the make invocation.
Removed code that looked at a pull down on the A0 line to determine what kind of LCD is used.

Removed the bootloader .lbm file for REVX build since it has to be compiled with the same LCD arguments as the main OS for it to work. See OpenRCbootloader git https://github.com/garyStofer/OpenRcBootloader